requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
autopep8=1.7.0=pypi_0
blas=2.115=openblas
blas-devel=3.9.0=15_linux64_openblas
bottleneck=1.3.4=py38hce1f21e_0
brotlipy=0.7.0=py38h0a891b7_1004
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.6.15=ha878542_0
certifi=2022.6.15=py38h578d9bd_0
cffi=1.15.1=py38h74dc2b5_0
charset-normalizer=2.1.0=pyhd8ed1ab_0
cloudpickle=2.1.0=pyhd8ed1ab_0
cryptography=37.0.1=py38h9ce1e76_0
cudatoolkit=11.7.0=hd8887f6_10
cudnn=8.4.1.50=hed8a83a_0
ffmpeg=4.3=hf484d3e_0
freetype=2.10.4=h0708190_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
gym=0.21.0=py38he5a9106_2
idna=3.3=pyhd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9e=h166bdaf_1
lame=3.100=h7f98852_1001
ld_impl_linux-64=2.38=h1181459_1
libblas=3.9.0=15_linux64_openblas
libcblas=3.9.0=15_linux64_openblas
libffi=3.4.2=h7f98852_5
libgcc-ng=12.1.0=h8d9b700_16
libgfortran-ng=12.1.0=h69a702a_16
libgfortran5=12.1.0=hdcd56e2_16
libgomp=12.1.0=h8d9b700_16
libiconv=1.17=h166bdaf_0
liblapack=3.9.0=15_linux64_openblas
liblapacke=3.9.0=15_linux64_openblas
libnsl=2.0.0=h7f98852_0
libopenblas=0.3.20=pthreads_h78a6416_0
libpng=1.6.37=h21135ba_2
libprotobuf=3.20.1=h6239696_0
libstdcxx-ng=12.1.0=ha89aaad_16
libtiff=4.0.10=hc3755c2_1005
libuuid=2.32.1=h7f98852_1000
libzlib=1.2.12=h166bdaf_2
llvm-openmp=14.0.4=he0ac6c6_0
lz4-c=1.9.3=h9c3ff4c_1
magma=2.5.4=h6103c52_2
mkl=2022.1.0=h84fe81f_915
mkl-service=2.4.0=py38h80f09db_0
mkl_fft=1.3.1=py38h887cb03_3
mkl_random=1.2.2=py38h5ca245f_1
nccl=2.13.4.1=h0800d71_0
ncurses=6.3=h5eee18b_3
nettle=3.6=he412f7d_0
ninja=1.11.0=h924138e_0
numexpr=2.7.3=py38h43a58ef_1
numpy=1.23.1=py38hf838250_0
numpy-base=1.23.1=py38h1e6e340_0
olefile=0.46=pyh9f0ad1d_1
openblas=0.3.20=pthreads_h320a7e8_0
openh264=2.1.1=h4ff587b_0
openssl=3.0.5=h166bdaf_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py38h295c915_0
pillow=6.2.1=py38h6b7be26_0
pip=22.1.2=py38h06a4308_0
pycodestyle=2.9.1=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py38h578d9bd_5
python=3.8.13=ha86cf86_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.12.1=cuda112py38habe9d5a_200
pytorch-gpu=1.12.1=cuda112py38h1894f8f_200
pytorch-mutex=1.0=cpu
pytz=2022.1=py38h06a4308_0
readline=8.1.2=h7f8727e_1
requests=2.28.1=pyhd8ed1ab_0
setuptools=61.2.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sleef=3.5.1=h9b69904_2
sqlite=3.39.2=h5082296_0
tbb=2021.5.0=h924138e_1
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pypi_0
torchaudio=0.12.1=py38_cpu
torchvision=0.13.1=py38_cpu
typing_extensions=4.3.0=pyha770c72_0
urllib3=1.26.11=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
zlib=1.2.12=h7f8727e_2
zstd=1.4.9=ha95c52a_0
